(*) Use floats instead of doubles.

+++ osuReplayAPI/LifeInfo.cs
@@ -4,6 +4,6 @@ namespace ReplayAPI
     public class LifeInfo
     {
         public int Time;
-        public double Percentage;
+        public float Percentage;
     }
 }
