+++ osuReplayAPI/ReplayInfo.cs
@@ -7,9 +7,9 @@ namespace ReplayAPI
         public Int64 TimeDiff;
         public int Time;
         [System.ComponentModel.DisplayName("Time In Seconds")]
-        public double TimeInSeconds { get { return Time/1000.0; } }
-        public double X { get; set; }
-        public double Y { get; set; }
+        public float TimeInSeconds { get { return Time/1000f; } }
+        public float X { get; set; }
+        public float Y { get; set; }
         public KeyData Keys { get; set; }
     }
 }
